Collect idleness measure in microseconds and plot charts

+++ Performance Tests/src/operations/locking_ca/structural_modification_ops_ca.cpp
@@ -43,15 +43,17 @@ int sb7::CAStructuralModification2::run(int tid) const {
     }
 
     auto *l = new lockObject (cpart->getLabellingId(), &cpart->criticalAncestors, 1);
-    pool.acquireLock(l, tid);
-    //// Between when the lock was created and the actual deletion happens,
-    /// If some other thread races to delete this object, Then, deleting will raise an exception.
-    /// We can solve this by taking two stage locks where the first lock is a read lock on the object which
-    /// prevents modifications and then we convert the read lock to a write lock.
-    if(cpart->hasLabel){
-        dataHolder->deleteCompositePart(cpart);
+    if(pool.acquireLock(l, tid)){
+        //// Between when the lock was created and the actual deletion happens,
+        /// If some other thread races to delete this object, Then, deleting will raise an exception.
+        /// We can solve this by taking two stage locks where the first lock is a read lock on the object which
+        /// prevents modifications and then we convert the read lock to a write lock.
+        if(cpart->hasLabel){
+            dataHolder->deleteCompositePart(cpart);
+        }
+        pool.releaseLock(l,tid);
     }
-    pool.releaseLock(l,tid);
+
     return 0;
 }
